requirements.txt
requests
pandas
scikit-learn
torch
python-dotenv
joblib
numpy<2.0.0
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
flask